// src/@types/student.d.ts
export interface Student {
    id: number
    first_name: string
    last_name: string
    dob: string
    gender: "female" | "male"
    marital_status: "single" | "divorced" | "married" | "widowed"
    email: string
    phone: string
    address: string
    nationality: string
    enrollment_date: string
    guardian_name: string
    guardian_contact: string
}